+++ changelog.md
@@ -12,6 +12,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityTablePanelPreferences bug fixed, column width preferences got lost for panels that were never initialized, now sets the column width as well as preferred width when applying preferences.
 - EntityTablePanelPreferences, auto-resize-mode preference added.
+- EntityTablePanelPreferences bug fixed, filter settings used instead of condition settings.
 
 ## 0.18.65
 ### is.codion.common.model